#13 EmailSignInActivity  네크워트 성공

+++ app/src/main/java/com/example/mindgarden/Network/POST/PostEmailSignInResponse.kt
@@ -11,5 +11,5 @@ data class EmailSignInData(
     val refreshToken:String,
     val email:String,
     val name:String,
-    val expires_in:Int
+    val expires_in:Long
 )
